requirements.txt
torch
torchvision
Pillow
numpy
opencv-python
scikit-image
